Pass a local variable to disable left input
Disable the left input (adding to users knowledge list) if there is no
current user. Only enabled the input when a valid user is logged in so
the correct associations can be made between their records and things
they know. Never disable the right input; always allow users to search
for others.