Move the user helper methods to the main controller
Move the 'user_signed_in?' and 'current_user' helper methods to the
Application Controller so that they can be used by both the controllers
and the views when determining content to render. Defining the
behaviors in the helper module made them only available to the views
when filling out the content to render.