Use local variables to render content lists
Generate the lists of things/users to render in the home controller and
pass the results in to the view as a local variable.

Results of the user search are generated from the string saved as a
session varaible set in the Things controller. The lists of things to
render users for is generated from the variable before it is cleared.